requirements.txt
setuptools
Flask
numpy
requests
python-dotenv
opencv-python
langchain-groq
langchain-community
uvicorn
fastapi
pydantic
langgraph
streamlit
langchain-tavily

